# perf(es): Add fast memory deallocator