--- a/crates/fastmem/Cargo.toml
+++ b/crates/fastmem/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+authors = ["강동윤 <kdy1997.dev@gmail.com>"]
+description = "Configurable utilities for fast memory operations"
+edition = "2021"
+license = "Apache-2.0"
+name = "fastmem"
+repository = "https://github.com/swc-project/swc.git"
+version = "0.1.0"
+
+[features]
+enable = ["rayon"]
+
+[dependencies]
+rayon = {version = "1", optional = true}