--- a/crates/fastmem/src/lib.rs
+++ b/crates/fastmem/src/lib.rs
@@ -1,0 +1,8 @@
+//! Utilities to unblock main thread.
+//!
+//! # Cargo Features
+//!
+//! ## `enable`
+//!
+//! This enables fast-mode of this crate.
+//!