--- a/crates/fastmem/src/lib.rs
+++ b/crates/fastmem/src/lib.rs
@@ -5,4 +5,10 @@
 //! ## `enable`
 //!
 //! This enables fast-mode of this crate.
-//!
+
+#![deny(clippy::all)]
+#![deny(unused)]
+
+pub use self::fast_drop::FastDrop;
+
+mod fast_drop;