--- a/crates/fastmem/src/lib.rs
+++ b/crates/fastmem/src/lib.rs
@@ -9,6 +9,6 @@
 #![deny(clippy::all)]
 #![deny(unused)]
 
-pub use self::fast_drop::FastDrop;
+pub use self::fast_drop::*;
 
 mod fast_drop;