--- a/crates/swc_node_base/Cargo.toml
+++ b/crates/swc_node_base/Cargo.toml
@@ -11,6 +11,7 @@ version = "0.5.1"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+fastmem = {version = "0.1.0", features = ["enable"]}
 
 [target.'cfg(not(all(target_os = "linux", target_arch = "aarch64", target_env = "musl")))'.dependencies]
 mimalloc-rust = "=0.1.5"