… Cargo.lock
… crates/swc_ecma_parser/scripts/instrument.sh