**Description:**



## Initial, main branch

```
running 8 tests
test angular       ... bench:   7,431,479 ns/iter (+/- 198,154) = 96 MB/s
test backbone      ... bench:   1,041,447 ns/iter (+/- 27,429) = 57 MB/s
test colors        ... bench:      14,379 ns/iter (+/- 257) = 80 MB/s
test jquery        ... bench:   5,720,166 ns/iter (+/- 141,980) = 46 MB/s
test jquery_mobile ... bench:   9,074,520 ns/iter (+/- 123,993) = 49 MB/s
test mootools      ... bench:   4,396,554 ns/iter (+/- 104,997) = 36 MB/s
test underscore    ... bench:     897,749 ns/iter (+/- 17,875) = 48 MB/s
test yui           ... bench:   4,301,158 ns/iter (+/- 98,899) = 78 MB/s
```